pyproject.toml
[tool.poetry]
name = "PROJECT_NAME"
version = "0.1.0"
description = ""
authors = ["Mohsen Padidar <mxpadidar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.2"
uvicorn = "^0.32.0"
sqlalchemy = "^2.0.36"
psycopg = "^3.2.3"
python-dotenv = "^1.0.1"
pyjwt = "^2.9.0"
bcrypt = "^4.2.0"
pydantic = { extras = ["email"], version = "^2.9.2" }
httpx = "^0.27.2"

[tool.poetry.group.dev.dependencies]
alembic = "^1.13.3"
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
pre-commit = "^4.0.1"
freezegun = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
pythonpath = ["./src"]
filterwarnings = "ignore::DeprecationWarning"
